[Calyx] [SCFToCalyx] Add collapse-control canonicalization pattern for IfOp, SeqOp and ParOp. (#1883)
Adds some canonicalization methods for SeqOp and ParOp to replace the pattern rewrite in SCFToCalyx. 
This is similar to the native compiler's Collapse Control pass. For the cells section, uses CHECK-DAG 
since we don't particularly care what order the cells are listed in. This was previously causing test failures, 
after adding the -canonicalize flag.