[FIRRTL] Improve readability of InferResets drive merging, NFC
Apply @youngar's suggestion in #1902 about improving the readability of
the part of `InferResets` which deals with merging of drives to reset
networks which have become unified. Rather than trying swap vectors out
and leave empty vectors in the map, just add the drives of the
non-leader network to the leader, and remove the non-leader's empty from
the map entirely.