[PyCDE] Use `System` to track raw CIRCT module ops
Keep all the raw CIRCT module ops in `pycde.System`. Enables the same PyCDE
modules to be used multiple Systems (theoretically, unsupported/untested for
now). It is also the only class which runs CIRCT passes, so it knows when to
invalidate the symbol cache. The symbol cache is the only data structure which
holds CIRCT module ops. Increases memory safety.