Object Model 2.0, Part 1: Read in OM 2.0 (#1915)
Add a new option to firtool, "--omir-file <file>", which will parse an
Object Model 2.0 JSON file into an OMIRAnnotation.  This only handles
parsing.  No scattering is performed.

Add one test that the "--omir-file <file>" option used to ingest an
Object Mdoel 2.0 JSON file works for firtool.  Check that the
information in the OMIR is preserved and is wrapped in an
OMIRAnnotation on the circuit.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>