[ExportVerilog] Start moving ExportVerilog to being a pass.
ExportVerilog was initially conceived as a translation that read the
IR and printed out verilog text.  However, it evolved to being a pass
that mutates the IR (e.g. its prepare pass).  As such, it makes more
sense to model it in CIRCT as a transformation: we want to be able to
run passes after it in the pass pipeline, e.g. to generate metadata.

This takes one step towards that, adding new entrypoints that make it
available in circt-opt.