[FIRRTL] Move more logic into ModuleLikeOpInterface
This moves support for port directions, names, types, and (port)
annotations into the `FModuleLike` operation interface.  This also
removes the `FuncLike` trait from `FExtModules` and `FModule`
operations.

Dropping support for `FuncLike` means that we are free to represent port
types as an array of types, instead of a functional type. This brings
some simplicity as port names and annotations were already arrays.

Another large difference is that we no longer support function argument
attributes.  There was a significant amount of complex code written to
support `argAttrs` and `portAnnotations` at the same time. Removing the
unused `argAttr` support simplified our code and IR printed format.

This work is not totaly complete, but I decided to cut it here before
the change became too large.  Some additional follow up work is:
1. Cleaning up the Direction attribute, utilizing new APInt 0 bit
   support.
2. Simplify FModuleOp's inputs TypeContraints so that the real
   verification is done by the FModuleLike verifier. Port existing tests
   over to the new verification logic.
3. Update `InstanceOp` and `MemOp` to use the same function names as
   `FModuleLike`, e.g. `getPortName()` instead of `getPortNameStr()`.