[FIRRTL] rename FModuleOp::getBodyBlock() to getBody() (#1940)
This change renames the `getBodyBlock()` function to `getBody()`, which matches
the CircuitOp API.  In a module there is only ever one region with one block in
it, and there is rarely a reason to get the Region. Naming this wasn't possible
before as `FunctionLike` was squatting on the `getBody()` name, which has
recently been dropped.