[FIRRTL][Calyx] Remove special handling of port direction for 0 ports
This change mostly affects FIRRTL, where Calyx has a small update to
make sure that some copied code matches identically.

This change removes the special handing for port directions when there
are zero ports.  Port directions are represented using an APInt as
a bitvector, and since APInts previously did not support 0 bit wide integers
there was some logic to handle this as a special case.  Since APInts
have been updated to allow 0 bit wide integers, this commit removes this
special handling.

In addition to this, some other cleanups included are:
- Moving the port direction logic from FIRRTLInterfaces to
  FIRRTLAttributes.  Since the port direction logic is not tied to the
  FModuleLike interface anymore, it can be used for other things like
  InstanceOps.
- Removing the `direction::attrKey` which was used as the attribute name
  for directions. Since this string is not a part of any OpInterface, it
  is unclear what code should rely on it. A future change should be made
  to stop using string literals and use
  `FModuleLike::getPortDirectionsAttrName()`.
- Fix printing of directions, and make them work with regular streams.
  This lets us use it in `printModuleSignature()`.
- Remove some verification from FExtModule which is already handled by
  FModuleLike.