[FIRRTL] Stop declaring FModuleLike attributes as arguments
This change removes the `portAnnotations` and `portNames`  attributes
from the FExtModule and FModule op's argument lists.  It is a little
unclear to me if this change is a good idea.  I noticed that some
attributes which are managed by OpInterfaces are not always declared as
ODS arguments, for example neither of these ops define `sym_name` from
`Symbol`. Here are the differences that this change makes as far as I
can tell:

1. Less obvious that these attributes exist on the operation (bad thing)
2. We don't have redundant `portAnnotations()` and
   `getPortAnnotations()` functions defined (good thing)
3. We don't have redundant verification of the attribute from the
   `AnnotationArrayAttr` type constraint and `FModuleLike` verifier
   (might be a good thing).
4. Does not affect the builder methods, as these ops do not use the
   generated builder methods. (no effect)

Given what I have seen in other interfaces, I think this is the way to
go.