[Simulator] Add support for HW aggregate operations (#1964)
Add support for HW aggregate operations in the LLHD simulator as preparation for #1653 and #1656 to avoid loosing simulator functionality.

Makes sure the endianess and bitcast semantics described in the rationale are followed. As HW uses big endian semantics and LLVM is machine dependent, the order of struct fields and array elements is reversed for little endian machines to match the in-memory layout. Big endian machines are not supported for now.

The lowering currently uses unpacked structs, however, HW structs are packed. This may lead to problems and should be fixed in a later PR (not included here as it adds a lot of noise).