[FIRRTL] Stop printing ExtModule port names with leading `%` character
This updates the signature printing for ExtModules to stop printing
argument names with a leading `%`.  This character indicates that a
value is an SSA value, when there is no attached Value or block under
the hood.

The name printing logic is using some code which should be upstreamed to
MLIR to only wrap the names in quotes if they contain problematic
characters. With this scheme, we will never have to print the portNames
element of the attribute dictionary for InstanceOps and FExtModules.
Regular FModules are unchanged and still print port names as regular
block arguments.

This change is in preparation of sharing the extmodule signature format
with the instance op when port names and directions are stored on the
instance operation.