[ExportVerilog] Move emitTypeDims/printPackedType into ModuleEmitter.
This makes way for dimensions that can refer to declarations within the
module, like parameters.

Note that this exposes some design problems with TypeAliasType.
I just xfailed a test for now (the expected diagnostics are emitted on
the wrong line now) where ExportVerilog is implementing verification of
these decls instead of the verifier.  I will tackle that as its own
separate workstream, since type parameters needs a sound design as well.