[FIRRTL] Use new AsmParser::parseOptionalAttribute function
This switches the version of `parseOptionalAttribute` used when parsing
port annotations in a module signature.  The old version of
`parseOptionalAttribute` would add the parsed attribute to a
NamedAttributeList, and we had to construct a dummy list to make this
work.  The new version doesn't require an attribute list to store the
attribute in.