[FIRRTL][LowerToHW] Drop annotations which have no effect
This change adds some more annotations to the list of annotations which
should be dropped in LowerToHW.

`ExtractGrandCentralAnnotation` is used by several grand central passes.

`SubCircuitsTargetDirectory` and `TestBenchDir` is an output directories
for unimplemented GrandCentral behaviour. They do not have any affect on
their own.

The `BlackBox` annotation is used to indicate which extmodule are
created from the BlackBoxReader pass and no pass is responsible for
removing it.

The `ElaborationArtefactsDirectory` specifies the output directory of a
not-implemented (and not planned) `ElaborationArtefacts` transform. In
order for this annotation to have any effect we would need a module
annotated with `ElaborationArtefactAnnotation`.

The `MetadataDirAnnotation` is used by a few passes and no
specific pass is responsible for removing it.