[FIRRTL][SV] Use `parseKeywordOrString` (#2013)
The `parseKeywordOrString` function was recently added to the MLIR
parser, and we can simplify some code by using it. We can remove the
functions `parsePort` from SV and `parseIdentifier` from FIRRTL.  These
functions all print a string as a keyword unless it contains illegal
characters, where it would print as a string.