[Python] Fix bindings after upstream changes to aggregate building.
With the new CMake functionality upstream, we need to set the
ENABLE_AGGREGATION option when we call add_mlir_library. The easiest
way to do this is to just rely on the upstream helper
add_mlir_public_c_api_library, which does this and other important
settings for us.

Also, ensure the MSFT dialect's CAPI marks all of the public functions
as being exported.