[FIRRTL] Support Prefixing of Grand Central Interfaces (#2020)
Change the Grand Central Views/Interfaces pass to apply an outer
module-level prefix (derived from a NestedPrefixModulesAnnotation that
was handled by the PrefixModules pass) and an inner prefix set by an
optional PrefixInterfacesAnnotation.  Modify RenameModules to set the
outer prefix by adding a "prefix" field to the circuit-level, outermost
AugmentedBundleType.

Supporting this, add a getter to retrieve the "prefix" field of an
AugmentedBundleType.

Add two tests: (1) test that PrefixModules correctly updates Grand
Central interface annotations and (2) test that the GrandCentral pass
correctly applies a prefix coming from PrefixModules and from a
PrefixInterfacesAnnotation.  As part of (2), ensure that the
PrefixInterfacesAnnotation is removed.

Update FIRRTL rationale document with information about how Grand
Central-generated interfaces can be prefixed.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>