[FIRRTL] Optimize foldFlow, NFC.
Instead of scanning the results of an instance to find the result
number (which is O(n) in # results) use OpResult to get it directly.

This is pretty hot in the profile because it is invoked frequently
by the verifier.  This change sped the full end to end (chirrtl to
verilog) time on a large testcase by 3%.