[ExportVerilog] Refactor isExpressionEmittedInline to be faster, NFC.
isExpressionUnableToInline is expensive: it walks the entire use-list
of an operation, which for some things can be a lot of stuff to sort
through.  However, most things that have more than one use can't
be inlined anyway.  Rearrange the control flow to only do the heavy
lifting in cases where it is needed.

This is a small but measurable speedup in ExportVerilog.