[FIRRTL] Fix SRAM instance paths in OMIR
Fix an issue in `EmitOMIR` where the instance paths of SRAMs would be
incorrectly rendered as `OMMemberReferenceTarget` instead of the correct
`OMMemberInstanceTarget`. This also includes figuring out the module
name that the final SRAM is going to have, which in the case of a
`firrtl.mem` is a bit tricky/fragile: the actual module name for the
memory is only generated after lowering to the HW dialect, at which
point the OMIR has long been emitted. This change "predicts" the name
that the memory is going to have by querying `getFirMemoryName()`, which
looks like it's going to be accurate in most cases we currently care
about.

In the long run, we'll want better ways to encode this formulation of an
instance path and ideally dealy it until `ExportVerilog` time.