[FIRRTL][GrandCentral] Make non-local paths local
This is a follow on to edd8956ec1c1, where we adapted to a new path
format for Grand Central augmented types. The problem with this new
format is that all created annotation targets are non-local, which is
which is not applying correctly using NLA. This change takes all given
non-local targets for augmented types and changes them to local targets,
without regard to whether these targets are totally equivalent.

The proper change is to either a) attach these as non-local annotations
and add support non-local annotations in Grand Central or b) have a
target canonicalizer which correctly makes these non-local annotations
local only when legal to do so.

Since this problem is currently blocking integration testing, I am
proposing this change for now.  I will be looking at moving the
GrandCentral annotation scattering into the newer
`firrtl-lower-annotation` pass, where hopefully we will accomplish
option `b`.