[FIRRTL] Extend when-encoded assert pattern detection in parser
Extend the FIR file parser's ability to detect when-encoded assertions.
In recent Chisel versions, these have slightly changed to now be of the
form:

    assert(clock, cond, enable, "msg")
    node T = eq(cond, UInt<1>(0))
    when T:
      printf(clock, enable, "Assertion failed: msg")

The previous implementation of the FIR parser only detects the following
pattern:

    when not_cond:
      printf(clock, enable, "Assertion failed: msg")
      stop(clock, enable, 1)

This fixes issues SiFive has seen on FIRRTL emitted from never versions
of Chisel.