[FIRRTL] Add RTL ports to OMIR data
Implement the functionality of the `SetOMIRPorts` in the Scala FIRRTL
compiler, which essentially extends all OMIR nodes that have a
`containingModule` field with a list of RTL ports of that module. There
are some quirks about `containingModule`, for example that it apparently
does not need to point to a module per se, that nee handling. Besides
that though, this is a fairly mechanical process.

This should enable IPXACT exporting based on the OMIR JSON that comes
out of the MLIR-based FIRRTL compiler.