[ExportVerilog] Emit HWModuleOp Comments
Add support for emitting comments associated with HWModuleOps during
ExportVerilog.  Use a simple line breaking algorithm to try to respect
the emittedLineLength option when reasonably possible.  Respect manual
line breaks that a user may have included in the comment.

Add tests that comments are emitted and that line breaking changes as
expected with changes to emittedLineLength.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>