[HandshakeToFIRRTL] Lower `handshake.instance` operations (#2067)
Since we're already relying on instantiating FIRRTL modules for the lowered handshake operations, instantiating an already lowered `handshake.func` module is straight-forward. The important part of this commit is that it ensures modules are lowered in post-order wrt. the instance graph (by iterating through the reversed topological order of the instance graph).