[HandshakeToFIRRTL] Add meaningful instance names
Previously, due to the lack of instance names, interpreting a waveform dump of a handshake circuit simulation was close to impossible given that all instance names were temporary names. This commit sets instance names as either the name of the `handshake.funcOp` submodule that is instantiated, or a shortened version of an operator submodule name.