[MSFT] Re-work ExportQuartusTcl to export into verbatim nodes.
This is the preferred way to output design collatoral
according. Closes https://github.com/llvm/circt/issues/2068.

The pass now populates a SymbolCache of the things we care about in
the export, namely instances. These instances have their symbol
attribute set, and the symbol is interpolated into the verbatim
output. Each verbatim node is assigned to a specific Tcl output file.

The old Python API to run the export is removed. In order to print Tcl
from PyCDE, the new pass is called followed by ExportVerilog, and the
Tcl output is extracted.