[FIRRTL] Accept `finalPath` as SRAM indicator for OMIR emission
Chisel has seen a change upstream where the `instancePath` field of
`OMSRAM` nodes is now called `finalPath`. There were additional changes
to the semantics of the field, all of which look like they only affect
the Scala side of things. From a CIRCT perspective, there's just a new
name for the path field that needs updating. To maintain compatibility
with not-yet-updated Chisel installations, the pass will accept the old
`instancePath` and the new `finalPath` fields.