[Calyx] Generate a hw.module.extern op for a calyx::ComponentOp
Returns a hw.module.extern operation describing the Verilog module which a ComponentOp eventually results in. This is useful for working with calyx modules in the rest of CIRCT at the hardware level, until we have support for further Calyx lowering inside CIRCT.

No tests are included in this commit; found it a bit excessive to create a test pass just for this. Tested locally, and I'm working on something to be pushed upstream which will use this functionality.