[FIRRTL] Fix MemOp port name emission
Fix an issue in the `FIREmitter` where a memory's ports, which are
represented by results of the operation, would not be assigned a name.
This would cause any use of the memory port later in the IR to trigger
an error instead of just using the `<memory_name>.<port_name>`
construct.

This fixes the issue and adds a regression test.