[FIRRTL] Fix FModuleOp::erasePorts failing on empty attr arrays
As an optimization, we use empty array attributes on module-like ops
where none of the ports have an attribute set. The code in `erasePorts`
assumes that all arrays have the same number of entries as the module
has ports.

This adds a special case that just passes through empty arrays (the
optimized case).