[ExportVerilog] Emit zero param as "0" not "64'd0" (#2076)
Change ExportVerilog emission of a zero-valued parameter to use "0" and
not "64'd0".  MLIR infrastructure will parse a zero as an APInt of type
i64.  This causes problem for emission of FIRRTL-parsed parameters which
need to show up in the output as "0" instead of "64'd0" which some tools
may reject.

This is special cased during Verilog emission to prevent having to
deviate from the seemingly canonical handling of a zero-valued APInt
being stored as an i64.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>