[FIRRTL]{Emitter] Add SeqMem and CombMem support
This adds support for emitting SeqMemOp, CombMemOp, and their memory
ports, MemoryPortOp and MemoryPortAccessOp.

The code for emitting memory port access is a little tricky because it
must reach backwards and find the original CHIRRTL memory declaration
and MemoryPort to get all the information needed to print.
Theoretically the port name could have been stored using `addValueName`,
but we already had the port op in hand so we could just ask it directly
for its name.

The code doesn't have error checking because the error cases are all
covered by the verifier.