[HIR] Added ForOp entry logic. Bugfixes.
BugFix - ArrayCreateOp takes input array of elements in reverse order.
The last element is treated as the index zero element. We had WRONGLY
assumed that the first element of the array input to the constructor of
ArrayCreateOp is the first element of the generated verilog array
(accessible at loc zero in verilog).

We dont support arith.constant for integers anymore. Use hw.constant.
For index type constants, we still support arith.constant.

SimplifyLoop pass is SimplifyCtrl pass.
Added ForOp entry logic in the above pass.

ForOp ub is not included in the range. This is to bring the behaviour of
ForOp same as it is in other dialects.

Bugfix - Due to incomplete operandMap, cloning sv dialect's AlwaysOp
resulted in wrong output since the SSA vars used inside the AlwaysOp
body were absent from the operandMap.