[FIRRTL] Add unknown width `mux(p, invalid, x) -> x` canonicalizer
This duplicates the functionality of a mux fold as a canonicalization to
support invalid values of unknown width.

A change to MLIR core disallows folds which change the return type of
the value used. When the invalid value is of unknown width, the mux's
return type will also be unknown width, but after the fold is performed
the type changes to the known width operand.

For example,
```
mux(p: uint<1>, invalid: uint, x: uint<2>) -> uint
```
canonicalizes to return `uint<2>`.