[FIRRTL] Add port symbols to module and extmodule ops
Extend the FIRRTL dialect such that the `FModuleOp` and `FExtModuleOp`
operations can carry an optional symbol for each of their ports. This is
done in a similar fashion as port names, directions, and annotations, in
that there is a separate array attribute which tracks the symbol for
each port. Since leaving null attributes in the IR is a bit yucky, the
absence of a symbol is encoded as an empty string.

This extends module-like port lists to support the following:

    firrtl.module @Foo(
      in %a: !firrtl.uint<1> sym @symbolA,
      out %b: !firrtl.uint<1> sym @symbolB
    )

The change to the IR itself is pretty trivial, but causes a whole bunch
of follow-up fixes in all the places that build or modify modules. These
places are surprisingly numerous and would benefit from some later
cleanup. @youngar has looked into moving the module-like signature into
a separate single attribute, which may go a long way towards making this
prettier.

This also adds tests to ensure port symbols are properly represented
within the FIRRTL dialect, they carry over into the HW dialect as
expected, and that using FIRRTL port symbols in `sv.verbatim` operations
has the intended Verilog output result.

This PR enables a whole bunch of follow-up work to make metadata and
OMIR emission cleaner and more robust in the presence of renaming.